created separate class file for vector class
